Refactor: Extract choosing and getting the media url from watch_media
