+++ src/cli/mod.rs
@@ -3,4 +3,5 @@ pub mod choose_lang;
 pub mod choose_media;
 pub mod choose_media_with_images;
 pub mod choose_season;
+pub mod get_media_url;
 pub mod get_medias;
